+++ src/singletons/WindowManager.cpp
@@ -168,6 +168,7 @@ WindowManager::WindowManager(const Paths &paths, Settings &settings,
 
     this->invalidateChannelViewBuffersListener.add(settings.alternateMessages);
     this->invalidateChannelViewBuffersListener.add(settings.separateMessages);
+    this->invalidateChannelViewBuffersListener.add(settings.fadeMessageHistory);
 
     this->repaintVisibleChatWidgetsListener.add(
         this->themes.repaintVisibleChatWidgets_);